collector/requirements.txt
kafka-python == 2.3.0